+++ src/index.jsx
@@ -1,12 +1,13 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.css';
-import App from './App';
 
 /* Bootstrap, Fontawesome, Jquery, Popper */
 import "bootstrap/dist/css/bootstrap.css";
 import '@fortawesome/fontawesome-free/js/all.js';
 import 'bootstrap/dist/js/bootstrap.bundle.min';
+
+import './index.css';
+import App from './App';
 
 var rootNode = document.getElementById('root');
 const root = createRoot(rootNode);